fix: 修复一些错误，主要是重构了各个命令类的代码。 :calling: :earth_africa:

+++ src/Commands/HashsetRemoveCommand.cs
@@ -2,7 +2,7 @@
  * Authors:
  *   钟峰(Popeye Zhong) <9555843@qq.com>
  *
- * Copyright (C) 2014-2016 Zongsoft Corporation <http://www.zongsoft.com>
+ * Copyright (C) 2014-2017 Zongsoft Corporation <http://www.zongsoft.com>
  *
  * This file is part of Zongsoft.Externals.Redis.
  *
@@ -49,17 +49,21 @@ namespace Zongsoft.Externals.Redis.Commands
 
 			var hashset = this.Redis.GetEntry<IRedisHashset>(context.Expression.Arguments[0]);
 
+			if(hashset == null)
+			{
+				context.Error.WriteLine($"The '{context.Expression.Arguments[0]}' hashset is not existed.");
+				return null;
+			}
+
 			if(context.Expression.Arguments.Length == 2)
-				hashset.Remove(context.Expression.Arguments[1]);
+				return hashset.Remove(context.Expression.Arguments[1]) ? 1 : 0;
 			else
 			{
 				var items = new string[context.Expression.Arguments.Length - 1];
 				Array.Copy(context.Expression.Arguments, 1, items, 0, items.Length);
 
-				hashset.RemoveRange(items);
+				return hashset.RemoveRange(items);
 			}
-
-			return null;
 		}
 		#endregion
 	}
